requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
dateparser==1.1.6
ffmpeg-python==0.2.0
filelock==3.9.0
frozenlist==1.3.3
future==0.18.3
googletrans-py==4.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.12.0
hyperframe==6.0.1
idna==3.4
more-itertools==9.0.0
multidict==6.0.4
numpy==1.24.2
openai-whisper==20230124
packaging==23.0
Pillow==9.4.0
py-cord==2.3.2
pycparser==2.21
PyNaCl==1.5.0
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
tokenizers==0.13.2
torch==1.13.1
tqdm==4.64.1
transformers==4.26.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
yarl==1.8.2
